Renamed all components to PascalCase Removed .vue extensions in Components imports Linter fixes
