+++ src/components/lib/CustomIcon.vue
