+++ src/views/Home.vue
@@ -8,10 +8,10 @@
 </template>
 
 <script>
-import Blurb from '@/components/blurb.vue';
-import About from '@/components/about.vue';
+import Blurb from '@/components/Blurb';
+import About from '@/components/About';
 import Projects from '@/components/Projects';
-import Skills from '@/components/Skills.vue';
+import Skills from '@/components/Skills';
 
 export default {
   name: 'Home',
